Prevent deleting processing image while attachment is processing
On destroying a model, Paperclip will delete any styles for which
exists? returns true. exists? checked the path, which would point to the
processing image while the attachment was processing when
processing_image_url is set. This leads to Paperclip deleting the
processing image, thinking it is the style for the attachment.

This patch overrides Paperclip's exists? method so that if the given
style is processing, exists? will always return false, thus preventing
Paperclip from trying to delete the processing image. This is not
backwards compatible as exists? previously returned true even if the
style was not yet processed.

This may still leave some orphaned files as if the model is destroyed
while the processing job is currently running, no styles will be
deleted, but some styles may have already been written to disk as the
job runs.